feat: enhance character management and session handling in timer-server
- Introduced character creation and login use cases, allowing users to create characters and log in to their accounts.
- Updated the repository structure to include character management functionalities.
- Enhanced session management to support character selection and updates.
- Added new gRPC protocol messages for character creation and updates.
- Refactored existing code to improve organization and maintainability, including updates to the PostgreSQL storage layer for character data.
- Updated the `.air.toml` for improved build commands and debugging capabilities.

+++ pkg/storage/postgres/repository/account.go
@@ -29,7 +29,7 @@ func (r *PostgresAccountRepo) FindByUsername(ctx context.Context, username strin
 	}
 
 	acc := &entity.Account{
-		ID:           uint32(dao.ID),
+		ID:           dao.ID,
 		Username:     dao.Username,
 		PasswordHash: dao.PasswordHash,
 		CreatedAt:    dao.CreatedAt.Time,
@@ -49,7 +49,7 @@ func (r *PostgresAccountRepo) Create(ctx context.Context, account *entity.Accoun
 		return err
 	}
 
-	account.ID = uint32(model.ID)
+	account.ID = model.ID
 	account.CreatedAt = model.CreatedAt.Time
 
 	return nil
